+++ CHANGELOG.md
@@ -15,11 +15,11 @@
  - Investigated performance of options for obtaining a checkout of a PR reference
    it was found that a local clone from the bare repo as a pre-step provided
    a meaningful performance improvement
+ - Replace the cloning actions with a subprocess call to a playbook
+   using the git module
 
 ### Roadmap
 
- - Replace the cloning actions with a subprocess call to a playbook
-   using the git module
  - Add a step that tests whether or not the ref (arbitrary ref) is present
     - if not present, the default branch is cloned into the tmp dir
       THIS IS HIGHLY CONTROVERSIAL